Split StatsScreen into a stats package
StatsScreen.kt was 1532 lines: the ViewModel, the screen, a streak
calendar, three metric cards, a pipeline chart and a long tail of small
formatting helpers. Continues the ui/word/ shape applied to Favorites,
Settings, Search and Developer.

  ui/stats/StatsViewModel.kt        UiState and ViewModel          221
  ui/stats/StatsScreen.kt           Screen, Content, header        214
  ui/stats/StatsCalendar.kt         streak card, month stepper,
                                    calendar grid and cells        437
  ui/stats/StatsCards.kt            effort, library and pipeline   636
  ui/stats/StatsFormatting.kt       stage/date/count helpers       174
  ui/stats/preview/StatsPreviews.kt sample data and four previews  209

Pure move. Diffing the moved bodies against the original shows the only
changes are twenty-one widenings, all forced by helpers that the cards
and calendar share across the new file boundaries — formatCount,
loadingAwareContentColor, StatsCard, CountText, stageLabel, stageColor
and the like. I took the list from the compiler rather than widening
pre-emptively, so nothing is exported that the split did not require.

StatsCountFormatterTest imports formatCountForLanguage from its new
package; it was already internal and needed no change beyond that.

Note that ui/stats has its own SectionHeader, distinct from the one in
ui/SettingsComponents.kt that Developer and Settings use. They stayed
separate here rather than being merged, which is a question for a later
pass over the shared components.

ui/ drops from 22 files to 21.
